requirements.txt
python-dotenv==1.0.1
python-docx==1.1.2
openai>=1.0.0
matplotlib>=3.8.0